Update npm replacements
