+++ server/embed/npm-replacements/abort-controller/polyfill.mjs
@@ -1,1 +1,1 @@
-export default null;
+export default {};
