+++ server/embed/npm-replacements/node-fetch.mjs
@@ -1,44 +1,18 @@
 export const { Blob, fetch, File, FormData, Headers, Request, Response, AbortController } = globalThis;
 export const AbortError = Error;
 export const FetchError = Error;
-
-const redirectStatus = /* @__PURE__ */ new Set([301, 302, 303, 307, 308]);
-export const isRedirect = (code) => redirectStatus.has(code);
+export const blobFrom = isNotSupport("blobFrom");
+export const blobFromSync = isNotSupport("blobFromSync");
+export const fileFrom = isNotSupport("fileFrom");
+export const fileFromSync = isNotSupport("fileFromSync");
+export const isRedirect = (code) => (code > 300 && code < 304) || (code > 306 && code < 309);
 
 fetch.isRedirect = isRedirect;
 fetch.Promise = globalThis.Promise;
 export default fetch;
 
-export async function blobFrom(path, type) {
-  if (typeof Deno !== "undefined") {
-    const file = await Deno.open(path);
-    const res = new Response(file.readable);
-    return new Blob([await res.blob()], { type });
-  }
-  throw new Error("blobFrom is not supported in browser");
-}
-
-export function blobFromSync(path, type) {
-  if (typeof Deno !== "undefined") {
-    const data = Deno.readFileSync(path);
-    return new Blob([data], { type });
-  }
-  throw new Error("blobFromSync is not supported in browser");
-}
-
-export async function fileFrom(path, type) {
-  if (typeof Deno !== "undefined") {
-    const file = await Deno.open(path);
-    const res = new Response(file.readable);
-    return new File([await res.blob()], path.split(/[\/\\]/).pop(), { type });
-  }
-  throw new Error("blobFrom is not supported in browser");
-}
-
-export function fileFromSync(path, type) {
-  if (typeof Deno !== "undefined") {
-    const data = Deno.readFileSync(path);
-    return new File([data], path.split(/[\/\\]/).pop(), { type });
-  }
-  throw new Error("blobFromSync is not supported in browser");
+function isNotSupport(fn) {
+  return () => {
+    throw new Error(`${fn} is not supported in browser`);
+  };
 }
